feat: ✨ configuración CORS y agregar enrutador de pruebas
