+++ app/domain/messages.py
@@ -1,0 +1,1 @@
+WELCOME_MESSAGE = "Sandy IA corriendo🚀"
