+++ app/services/application/status_service.py
@@ -1,0 +1,4 @@
+from app.domain.messages import WELCOME_MESSAGE
+
+def get_root_message():
+    return WELCOME_MESSAGE
